+++ tests/Parser/ExplainStatementTest.php
@@ -24,6 +24,14 @@ class ExplainStatementTest extends TestCase
         return [
             ['parser/parseExplain'],
             ['parser/parseExplain1'],
+            ['parser/parseExplain2'],
+            ['parser/parseExplain3'],
+            ['parser/parseExplain4'],
+            ['parser/parseExplainErr'],
+            ['parser/parseExplainErr1'],
+            ['parser/parseExplainErr2'],
+            ['parser/parseExplainErr3'],
+            ['parser/parseExplainErr4'],
         ];
     }
 }